requirements.txt
python-telegram-bot>=20.0
sqlalchemy
gofile
python-dotenv
